feat: added material calculation and more code separation
